+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### [12.130.0](https://kaos.sh/ek/12.130.0)
 
-- `[knf/validators]` Added validators `Set`, `SetToAny`, and `SetToAnyIgnoreCase`
-- `[knf/validators]` Validators `Empty` and `NotContains` marked as deprecated
+- `[knf/validators]` Added validators `Set`, `SetToAny`, `SetToAnyIgnoreCase`, `LenLess`, `LenGreater`, and `LenEquals`
+- `[knf/validators]` Validators `Empty`, `NotContains`, and `NotLen` marked as deprecated
 - `[knf/validators]` Improved validators input validation
 - `[knf/validators]` Code refactoring